serialised input as default feed id
